+++ docs/changelog.md
@@ -1,5 +1,6 @@
 - Fix debug mode of the $transform operation (#572)
 - GUI: ask the engine to analyze Bundle documents for possible validation profiles (#355)
+- GUI: fix the XML syntax highlighting
 
 2026/08/31 Release 4.1.14